feat: Added Model, Factories, and Seeder for Product, Category, Order, and User

+++ database/migrations/2023_12_26_174833_create_products_table.php
@@ -12,7 +12,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->decimal('price', 8, 2);
+            $table->bigInteger('price');
             $table->timestamps();
         });
     }
